+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Web UI version numbers should always match the corresponding version of LBRY App
   * New welcome flow for new users
   * Redesigned UI for Discover
   * Handle more of price calculations at the daemon layer to improve page load time
+  * Add special support for building channel claims in lbryuri module
 
 ### Changed
   * Update process now easier and more reliable